+++ usr/share/archlinux-tweak-tool/packages_gui.py
@@ -4,7 +4,6 @@
 
 import functools
 
-import functions as fn
 import packages
 from packages import Packages